app1/requirements.txt
# Application 1: AI-Powered Forex Strategy Generator
# Core ML & Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0

# Financial Data
pandas-ta==0.3.14b0
ta-lib>=0.4.25
yfinance>=0.2.18

# News & Sentiment
newspaper3k>=0.2.8
textblob>=0.17.1
vaderSentiment>=3.3.2
requests>=2.31.0
httpx>=0.24.0
feedparser>=6.0.10
tweepy>=4.14.0

# BERT & Sentiment Analysis
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0
datasets>=2.12.0
accelerate>=0.20.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
redis>=4.6.0
influxdb-client>=1.36.0

# API
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiohttp>=3.8.0
websockets>=11.0.0

# Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Logging
loguru>=0.7.0

# Strategy Validation & Backtesting
backtrader>=1.9.78.123
asyncpg>=0.28.0
